Fix some formatting and naming
Finally, an overview of the global changes will be provided, providing a solution with a minimum amount of changes. Try to put all the changes in one file. For such different branches, we only need to add this file. Currently in the development stage, some modifications are reflected in specific pages. This is only temporary.